Se arreglo un bug en la obtencion del Bearer token
Se modifico el server controller para que solo invalide un token
particular y no todos los tokens pendientes del server
